+++ noodles-bam/src/record/codec/encoder/name.rs
@@ -1,21 +1,38 @@
-use bytes::BufMut;
-use noodles_sam::{self as sam, record::Name};
+use std::io;
 
-pub fn put_name<B>(dst: &mut B, name: Option<&Name>)
+use bytes::BufMut;
+use noodles_sam::alignment::record_buf::Name;
+
+const MAX_LENGTH: usize = 254;
+const MISSING: &[u8] = b"*";
+
+pub fn put_name<B>(dst: &mut B, name: Option<&Name>) -> io::Result<()>
 where
     B: BufMut,
 {
-    use sam::record::name::MISSING;
-
     const NUL: u8 = 0x00;
 
     if let Some(name) = name {
-        dst.put(name.as_ref());
+        let buf = name.as_ref();
+
+        if !is_valid(buf) {
+            return Err(io::Error::from(io::ErrorKind::InvalidInput));
+        }
+
+        dst.put(buf);
     } else {
         dst.put(MISSING);
     }
 
     dst.put_u8(NUL);
+
+    Ok(())
+}
+
+fn is_valid(buf: &[u8]) -> bool {
+    (1..=MAX_LENGTH).contains(&buf.len())
+        && buf != MISSING
+        && buf.iter().all(|&b| b.is_ascii_graphic() && b != b'@')
 }
 
 #[cfg(test)]
@@ -23,19 +40,39 @@ mod tests {
     use super::*;
 
     #[test]
-    fn test_put_name() -> Result<(), sam::record::name::ParseError> {
-        fn t(buf: &mut Vec<u8>, name: Option<&Name>, expected: &[u8]) {
+    fn test_put_name() -> io::Result<()> {
+        fn t(buf: &mut Vec<u8>, name: Option<&Name>, expected: &[u8]) -> io::Result<()> {
             buf.clear();
-            put_name(buf, name);
+            put_name(buf, name)?;
             assert_eq!(buf, expected);
+            Ok(())
         }
 
         let mut buf = Vec::new();
 
-        t(&mut buf, None, &[b'*', 0x00]);
-        t(&mut buf, Some(&"r".parse()?), &[b'r', 0x00]);
-        t(&mut buf, Some(&"r0".parse()?), &[b'r', b'0', 0x00]);
+        t(&mut buf, None, &[b'*', 0x00])?;
+        t(&mut buf, Some(&Name::from(b"r0")), &[b'r', b'0', 0x00])?;
 
         Ok(())
+    }
+
+    #[test]
+    fn test_put_name_with_invalid_name() {
+        fn t(raw_name: &[u8]) {
+            let mut buf = Vec::new();
+
+            assert!(matches!(
+                put_name(&mut buf, Some(&Name::from(raw_name))),
+                Err(e) if e.kind() == io::ErrorKind::InvalidInput
+            ));
+        }
+
+        t(b"");
+        t(b"*");
+        t(b"r 0");
+        t(b"@r0");
+
+        let s = vec![b'n'; MAX_LENGTH + 1];
+        t(&s);
     }
 }